# internal/handlers/userHandler.go
// Package handlers provides HTTP request handler functions for a web service written in Go
package handlers

import (
	"context"
	"fmt"
	"io"
	"net/http"
	"os"
	"path/filepath"

	"github.com/eugenshima/myapp/internal/model"

	"github.com/go-playground/validator"
	"github.com/google/uuid"
	"github.com/labstack/echo/v4"
	"github.com/sirupsen/logrus"
)

// UserHandlerImpl struct represents a user handler implementation
type UserHandlerImpl struct {
	srv UserService
}

// NewUserHandlerImpl creates a new Handler
func NewUserHandlerImpl(srv UserService) *UserHandlerImpl {
	return &UserHandlerImpl{
		srv: srv,
	}
}

// UserService interface implementation
type UserService interface {
	GenerateTokens(ctx context.Context, login, password string) (string, string, error)
	Signup(ctx context.Context, entity *model.User) error
	RefreshTokenPair(ctx context.Context, accessToken string, refreshToken string, id uuid.UUID) (string, string, error)
	GetAll(ctx context.Context) ([]*model.User, error)
}

// Login receives a GET request from client and returns a user(if exists)
// @Summary Login user
// @tags authentication methods
// @Description Logs in a user and returns access and refresh tokens
// @Accept json
// @Produce json
// @Param input body model.Login true "Login details"
// @Success 200 {object} map[string]interface{} " Generating access and refresh tokens"
// @Failure 404 {string} string "Error message"
// @Router /api/user/login [post]
func (handler *UserHandlerImpl) Login(c echo.Context) error {
	input := model.Login{}
	err := c.Bind(&input)
	if err != nil {
		logrus.Errorf("Error in userHandler: %v", err)
		return c.String(http.StatusNotFound, fmt.Sprintf("Error in userHandler: %v", err))
	}

	accessToken, refreshToken, err := handler.srv.GenerateTokens(c.Request().Context(), input.Login, input.Password)
	if err != nil {
		logrus.Errorf("error Generating JWT token %v", err)
		return c.String(http.StatusNotFound, fmt.Sprintf("Error in userHandler: %v", err))
	}
	response := map[string]interface{}{
		"access_token":  accessToken,
		"refresh_token": refreshToken,
	}
	return c.JSON(http.StatusOK, response)
}

// Signup receives a POST request from client to sign up a user
// @Summary Sign up user
// @tags authentication methods
// @Description Registers a new user
// @Accept json
// @Produce json
// @Param reqBody body model.Signup true "Signup details"
// @Success 200 {string} string "User created"
// @Failure 400 {string} string "Error message"
// @Failure 500 {string} string "Internal server error"
// @Router /api/user/signup [post]
func (handler *UserHandlerImpl) Signup(c echo.Context) error {
	reqBody := model.Signup{}
	err := c.Bind(&reqBody)
	if err != nil {
		logrus.Errorf("Error in userHandler: %v", err)
		return c.String(http.StatusNotFound, fmt.Sprintf("Error in userHandler: %v", err))
	}
	entity := &model.User{
		ID:       uuid.New(),
		Login:    reqBody.Login,
		Password: []byte(reqBody.Password),
		Role:     reqBody.Role,
	}

	validate := validator.New()
	if err = validate.Struct(entity); err != nil {
		logrus.Errorf("error in handler: %v", err)
		str := fmt.Sprintf("Error in handler: %v", err)
		return c.String(http.StatusBadRequest, str)
	}

	err = handler.srv.Signup(c.Request().Context(), entity)
	if err != nil {
		logrus.Errorf("error calling Signup method: %v", err)
		return c.String(http.StatusInternalServerError, fmt.Sprintf("Error in userHandler: %v", err))
	}

	return c.JSON(http.StatusOK, "Created")
}

// GetAll receives a POST request from client for getting all entities from the server
func (handler *UserHandlerImpl) GetAll(c echo.Context) error {
	results, err := handler.srv.GetAll(c.Request().Context())
	if err != nil {
		logrus.Errorf("error calling GetAll method: %v", err)
		return c.String(http.StatusNotFound, fmt.Sprintf("Error in userHandler: %v", err))
	}
	return c.JSON(http.StatusOK, results)
}

// RefreshTokenPair receives a POST request from client for refreshing a token pair
// @Summary Refresh token pair
// @tags authentication methods
// @Description Refreshes an access token and a refresh token
// @Accept json
// @Produce json
// @Param id path string true "ID of the user"
// @Param reqBody body model.Tokens true "Token pair details"
// @Success 200 {object} map[string]interface{} "Refreshed token pair"
// @Failure 400 {string} string "Bad request"
// @Failure 404 {string} string "Error message"
// @Router /api/user/refresh/{id} [post]
func (handler *UserHandlerImpl) RefreshTokenPair(c echo.Context) error {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		logrus.Errorf("error calling RefreshTokenPair method: %v", err)
		return c.String(http.StatusBadRequest, fmt.Sprintf("Error in userHandler: %v", err))
	}
	reqBody := model.Tokens{}
	err = c.Bind(&reqBody)
	if err != nil {
		logrus.Errorf("error calling RefreshTokenPair method: %v", err)
		return c.String(http.StatusNotFound, fmt.Sprintf("Error in userHandler: %v", err))
	}

	accessToken, refreshToken, err := handler.srv.RefreshTokenPair(c.Request().Context(), reqBody.AccessToken, reqBody.RefreshToken, id)
	if err != nil {
		logrus.Errorf("error calling RefreshTokenPair method: %v", err)
		return c.String(http.StatusNotFound, fmt.Sprintf("Error in userHandler: %v", err))
	}
	response := map[string]interface{}{
		"access_token":  accessToken,
		"refresh_token": refreshToken,
	}
	return c.JSON(http.StatusOK, response)
}

// image struct for downlodaing images from internet
type image struct {
	Filename string `json:"filename"`
	Url      string `json:"url"`
}

// GetImage returns an image from database
// @Summary Get image by name
// @Security ApiKeyAuth
// @tags download/upload images
// @Description Retrieves an image by name
// @Produce octet-stream
// @Param name path string true "Name of the image"
// @Success 200 {file} file "Image file"
// @Failure 404 {string} string "Image not found"
// @Router /api/image/get/{name} [get]
func (handler *UserHandlerImpl) GetImage(c echo.Context) error {
	image := c.Param("name")
	filePath := "/home/yauhenishymanski/MyProject/myapp/internal/images/" + image
	return c.Inline(filePath, image)
}

// SetImage saves the image from the internet
// @Summary Set image
// @Security ApiKeyAuth
// @tags download/upload images
// @Description Sets an image from the provided URL
// @Accept json
// @Produce plain
// @Param img body image true "Image details"
// @Success 200 {string} string "Image has been set"
// @Failure 400 {string} string "Bad request"
// @Failure 404 {string} string "Error message"
// @Router /api/image/set [post]
func (handler *UserHandlerImpl) SetImage(c echo.Context) error {
	img := image{}
	err := c.Bind(&img)
	if err != nil {
		logrus.Errorf("error in handler: %v", err)
		return c.String(http.StatusBadRequest, fmt.Sprintf("Error in handler: %v", err))
	}
	response, err := http.Get(img.Url)
	if err != nil {
		logrus.Errorf("error in handler: %v", err)
		return c.String(http.StatusBadRequest, fmt.Sprintf("Error in handler: %v", err))
	}
	defer response.Body.Close()

	filename := filepath.Join("/home/yauhenishymanski/MyProject/myapp/internal/images", img.Filename)
	file, err := os.Create(filename)
	if err != nil {
		logrus.Errorf("error in handler: %v", err)
		return c.String(http.StatusBadRequest, fmt.Sprintf("Error in handler: %v", err))
	}
	defer file.Close()

	_, err = io.Copy(file, response.Body)
	if err != nil {
		logrus.Errorf("error in handler: %v", err)
		return c.String(http.StatusBadRequest, fmt.Sprintf("Error in handler: %v", err))
	}
	return c.String(http.StatusOK, "image has been set")
}